+++ backend/druks/accounts/constants.py
@@ -1,0 +1,3 @@
+# Owns every run nobody asked for: crons, background work. Seeded at first
+# start; no provider email ever collides with it.
+SYSTEM_ACCOUNT_ID = "system"
